+++ src/iop/colorreconstruction.c
@@ -667,7 +667,7 @@ void process(struct dt_iop_module_t *self, dt_dev_pixelpipe_iop_t *piece, const 
 error:
   dt_control_log(_("module `color reconstruction' failed"));
   dt_iop_colorreconstruct_bilateral_free(b);
-  memcpy(ovoid, ivoid, (size_t)sizeof(float) * piece->colors * roi_out->width * roi_out->height);
+  memcpy(ovoid, ivoid, sizeof(float) * piece->colors * roi_out->width * roi_out->height);
 }
 
 #ifdef HAVE_OPENCL
